requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --strip-extras
#
numpy==2.1.2
    # via pandas
pandas==2.2.3
    # via -r requirements.in
patool==3.1.0
    # via -r requirements.in
pypdf2==3.0.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
six==1.16.0
    # via python-dateutil
tzdata==2024.2
    # via pandas
